+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # QOwnNotes Changelog
 
+## 1.0.6
+- you can now enter **network proxy** settings in the new network settings 
+  tab to make QOwnNotes work over network proxies
+  (for [Issue #79](https://github.com/pbek/QOwnNotes/issues/79))
+
 ## 1.0.5
 - Google Analytics got removed in favor of Piwik
 - code blocks will now wrap around selected text when inserted